movie-recommendation-chatbot/pyproject.toml
[project]
name = "movie-recommendation-chatbot"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.6",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
]
